Turned on (and resolved) more compiler warnings.

+++ CoreLib/Parsers/PCKXMLParserDelegate.m
@@ -25,8 +25,8 @@
             const char * startAttributeValue = attributes[i + ATTRIBUTE_START_OFFSET];
             const char * endAttributeValue = attributes[i + ATTRIBUTE_END_OFFSET];
 
-            int attributeBytes = (unsigned long)endAttributeValue - (unsigned long)startAttributeValue;
-            char * attributeValue = (char *)malloc(attributeBytes + 1);
+            ptrdiff_t attributeBytes = endAttributeValue - startAttributeValue;
+            char * attributeValue = (char *)malloc((size_t)attributeBytes + 1);
             attributeValue[attributeBytes] = 0;
             strncpy(attributeValue, startAttributeValue, attributeBytes);
             [attributesDictionary setValue:[NSString stringWithCString:attributeValue encoding:NSUTF8StringEncoding] forKey:attributeName];
